Fix for feature id missing in MVT asset layers
This is a fix (potentially temporary) for broken auto-highlight and selection for asset layers.
Not clear right now if the problem is caused by an update to deck.gl removing the feature IDs,
or if the MVT features actually don't contain the feature id (but rather have an id property),
and the deck.gl update simply made that visible.